+++ src/AzureCloud/AzureResourceOptions.cs
@@ -5,8 +5,17 @@ using Microsoft.Extensions.Configuration;
 
 namespace Squadron.AzureCloud
 {
+    /// <summary>
+    /// Base options to use with Azure resources
+    /// </summary>
+    /// <seealso cref="Squadron.AzureCloud.IAzureResourceConfigurationProvider" />
     public class AzureResourceOptions : IAzureResourceConfigurationProvider
     {
+
+        /// <summary>
+        /// Gets the azure configuration using the .NET configuration system
+        /// </summary>
+        /// <returns></returns>
         public AzureResourceConfiguration GetAzureConfiguration()
         {
             IConfigurationRoot configuration = new ConfigurationBuilder()
