+++ src/AzureCloudServiceBus/AzureCloudServiceBusOptions.cs
@@ -6,8 +6,16 @@ using Squadron.AzureCloud;
 
 namespace Squadron
 {
+    /// <summary>
+    /// ServiceBus resources options
+    /// </summary>
+    /// <seealso cref="Squadron.AzureCloud.AzureResourceOptions" />
     public abstract class AzureCloudServiceBusOptions : AzureResourceOptions
     {
+        /// <summary>
+        /// Configures the ServiceBus
+        /// </summary>
+        /// <param name="builder">The builder.</param>
         public abstract void Configure(ServiceBusOptionsBuilder builder);
     }
 }
