+++ src/AzureCloudServiceBus/Model/ServiceBusTopicModel.cs
@@ -2,6 +2,9 @@ using System.Collections.Generic;
 
 namespace Squadron
 {
+    /// <summary>
+    /// ServiceBusTopic
+    /// </summary>
     public class ServiceBusTopicModel
     {
         internal ServiceBusTopicModel()
@@ -9,35 +12,38 @@ namespace Squadron
 
         }
 
+        /// <summary>
+        /// Initializes a new instance of the <see cref="ServiceBusTopicModel"/> class.
+        /// </summary>
+        /// <param name="name">The name.</param>
         public ServiceBusTopicModel(string name)
         {
             Name = name;
         }
 
+        /// <summary>
+        /// Gets or sets the name.
+        /// </summary>
+        /// <value>
+        /// The name.
+        /// </value>
         public string Name { get; set; }
 
+        /// <summary>
+        /// Gets or sets the generated name
+        /// </summary>
+        /// <value>
+        /// The name of the created.
+        /// </value>
         internal string CreatedName { get; set; }
 
+        /// <summary>
+        /// Gets or sets the subscriptions.
+        /// </summary>
+        /// <value>
+        /// The subscriptions.
+        /// </value>
         public List<ServiceBusSubscriptionModel> Subscriptions { get; set; }
             = new List<ServiceBusSubscriptionModel>();
     }
-
-    public class ServiceBusQueueModel
-    {
-        public ServiceBusQueueModel(string name)
-        {
-            Name = name;
-        }
-        public string Name { get; set; }
-        internal string CreatedName { get; set; }
-
-    }
-
-    public class ServiceBusSubscriptionModel
-    {
-        public string Name { get; set; }
-
-        public string SqlFilter { get; set; }
-    }
-
 }
