+++ src/AzureStorage/Blob/AzureStorageBlobDefaultOptions.cs
@@ -1,7 +1,14 @@
-﻿namespace Squadron
+namespace Squadron
 {
+    /// <summary>
+    /// Default AzureStorage blob resource options
+    /// </summary>
     public class AzureStorageBlobDefaultOptions : ContainerResourceOptions
     {
+        /// <summary>
+        /// Configure resource options
+        /// </summary>
+        /// <param name="builder"></param>
         public override void Configure(ContainerResourceBuilder builder)
         {
             var name = "azurite_blob";
