+++ src/Core/ContainerInstance.cs
@@ -1,18 +1,61 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 
 namespace Squadron
 {
+    /// <summary>Respresents a created docker container</summary>
     public class ContainerInstance
     {
+        /// <summary>
+        /// Gets or sets the identifier.
+        /// </summary>
+        /// <value>
+        /// The identifier.
+        /// </value>
         public string Id { get; set; }
 
+
+        /// <summary>
+        /// Gets or sets the address.
+        /// </summary>
+        /// <value>
+        /// The address.
+        /// </value>
         public string Address { get; set; }
 
+
+        /// <summary>
+        /// Gets or sets the ip address.
+        /// </summary>
+        /// <value>
+        /// The ip address.
+        /// </value>
         public string IpAddress { get; set; }
 
+
+        /// <summary>
+        /// Gets or sets the host port.
+        /// </summary>
+        /// <value>
+        /// The host port.
+        /// </value>
         public int HostPort { get; set; }
+
+
+        /// <summary>
+        /// Gets a value indicating whether this instance is running.
+        /// </summary>
+        /// <value>
+        ///   <c>true</c> if this instance is running; otherwise, <c>false</c>.
+        /// </value>
         public bool IsRunning { get; internal set; }
 
+
+        /// <summary>
+        /// Gets or sets the container logs.
+        /// </summary>
+        /// <value>
+        /// The logs.
+        /// </value>
         public IList<string> Logs { get; set; } = new List<string>();
     }
 }
