+++ src/Core/ContainerResource.cs
@@ -4,15 +4,37 @@ using System.Threading.Tasks;
 
 namespace Squadron
 {
+
+    /// <summary>
+    /// Base resource for container based resources
+    /// </summary>
+    /// <typeparam name="TOptions">The type of the options.</typeparam>
     public class ContainerResource<TOptions>
          where TOptions : ContainerResourceOptions, new()
     {
+        /// <summary>
+        /// Gets or sets the continer settings.
+        /// </summary>
+        /// <value>
+        /// The settings.
+        /// </value>
         protected ContainerResourceSettings Settings { get; set; }
 
+
+        /// <summary>
+        /// The container manager
+        /// </summary>
         protected IDockerContainerManager Manager = null;
 
+
+        /// <summary>
+        /// The container initializer
+        /// </summary>
         protected ContainerInitializer Initializer = null;
 
+        /// <summary>
+        /// Initializes the resources
+        /// </summary>
         public async virtual Task InitializeAsync()
         {
             var options = new TOptions();
@@ -25,6 +47,9 @@ namespace Squadron
             await Manager.CreateAndStartContainerAsync();
         }
 
+        /// <summary>
+        /// Cleans up the resource
+        /// </summary>
         public async Task DisposeAsync()
         {
             try
