+++ src/Redis.Tests/Redis.Tests.csproj
@@ -25,4 +25,10 @@
     <ProjectReference Include="..\Redis\Redis.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <None Update="xunit.runner.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
 </Project>
