+++ src/Redis.Tests/xunit.runner.json
@@ -1,0 +1,4 @@
+﻿{
+  "appDomain": "denied",
+  "parallelizeAssembly": true
+}
